feat: 시작 화면 구현
