+++ backend/seats/views.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def home(request):
+    return render(request, "seats/home.html")
