+++ frontend/.gitignore
@@ -1,0 +1,102 @@
+### Django ###
+*.log
+*.pot
+*.pyc
+__pycache__/
+local_settings.py
+db.sqlite3
+db.sqlite3-journal
+media/
+
+# Python / Django
+*.py[cod]
+*$py.class
+*.so
+
+# Environments
+.env
+.venv/
+env/
+ENV/
+env.bak/
+venv.bak/
+
+# Packaging
+.Python
+build/
+dist/
+.eggs/
+*.egg-info/
+*.egg
+MANIFEST
+
+# Testing / Coverage
+htmlcov/
+.tox/
+.nox/
+.coverage*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+.pytest_cache/
+cover/
+.mypy_cache/
+.dmypy.json
+dmypy.json
+
+# Others
+*.mo
+*.manifest
+*.spec
+pip-log.txt
+pip-delete-this-directory.txt
+docs/_build/
+.pybuilder/
+target/
+.ipynb_checkpoints
+instance/
+.scrapy
+/site
+cython_debug/
+
+### Next.js ###
+node_modules/
+.pnp
+.pnp.js
+coverage/
+.next/
+out/
+build/
+.DS_Store
+*.pem
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+.pnpm-debug.log*
+.env*.local
+.vercel
+*.tsbuildinfo
+next-env.d.ts
+
+### Virtualenv ###
+[Bb]in/
+[Ii]nclude/
+[Ll]ib/
+[Ll]ib64/
+[Ll]ocal/
+[Ss]cripts/
+pyvenv.cfg
+pip-selfcheck.json
+
+### VSCode ###
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+!.vscode/*.code-snippets
+.history/
+*.vsix
+.ionide
+
